Added financial transaction to NC
