+++ app/lib/submission_builder/ty2024/states/nc/nc_return_xml.rb
@@ -18,6 +18,12 @@ module SubmissionBuilder
             "NCIndividual2023v1.0"
           end
 
+          def build_state_specific_tags(document)
+            if !@submission.data_source.routing_number.nil? && !@submission.data_source.account_number.nil?
+              document.at("ReturnState").add_child(financial_transaction)
+            end
+          end
+
           def schema_file
             SchemaFileLoader.load_file("us_states", "unpacked", "NCIndividual2023v1.0", "NCIndividual", "IndividualReturnNCD400.xsd")
           end
@@ -42,6 +48,17 @@ module SubmissionBuilder
           def adjustments_present?
             calculated_fields = @submission.data_source.tax_calculator.calculate
             (calculated_fields[:NCD400_LINE_9]).positive? # TODO Also needs to check if Line 7 is positive but not in scope yet
+          end
+
+          def financial_transaction
+            calculator = @submission.data_source.tax_calculator
+            calculator.calculate
+
+            FinancialTransaction.build(
+              @submission,
+              validate: false,
+              kwargs: { refund_amount: calculator.refund_or_owed_amount }
+            ).document.at("*")
           end
         end
       end
