+++ db/migrate/20241112175305_change_nc_routing_number_to_string.rb
@@ -1,0 +1,6 @@
+class ChangeNcRoutingNumberToString < ActiveRecord::Migration[7.1]
+  def change
+    # This column cannot be written to as it is, due to it giving an encryption error, should always be empty
+    safety_assured { change_column :state_file_nc_intakes, :routing_number, :string }
+  end
+end
