+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.1].define(version: 2024_11_01_161438) do
+ActiveRecord::Schema[7.1].define(version: 2024_11_12_175305) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "citext"
   enable_extension "plpgsql"
@@ -2003,7 +2003,7 @@ ActiveRecord::Schema[7.1].define(version: 2024_11_01_161438) do
     t.jsonb "raw_direct_file_intake_data"
     t.string "referrer"
     t.string "residence_county"
-    t.integer "routing_number"
+    t.string "routing_number"
     t.decimal "sales_use_tax", precision: 12, scale: 2
     t.integer "sales_use_tax_calculation_method", default: 0, null: false
     t.integer "sign_in_count", default: 0, null: false
